LayoutTests/fast/parser/block-nesting-cap-table-expected.txt

Test that the HTML parser does not allow the nesting depth of "block-level" elements to exceed 4096 when using nested table tag

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS t1024.parentNode === td1022 is true
PASS t1023.parentNode === td1022 is true
PASS t1024.previousSibling === t1023 is true
PASS successfullyParsed is true

TEST COMPLETE
